requirements.txt
aws-cdk-lib==2.100.0
boto3==1.34.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
cryptography==41.0.8